Added the SQL from this course
AdventureWorks uses the online database to create procedures and query the database.
In Library Management SQL, I create the database from scratch and populate it. Then I create procedures to query what is in the database.
In the Video Games SQL I create a smaller video game company library with details about them and their games. Creating the database, and stored procedures as well.